backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.3.2
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
urllib3==1.26.9
